requirements.txt
flask==3.0.2
python-dotenv==1.0.1
flask-sock==0.7.0
psutil==5.9.8
simple-websocket==1.0.0  # Required by flask-sock
boto3==1.35.95